Initial work towards wasm support
This has several changes: first, easyloggingpp is vendored (and a
trivially patched version is used for now to add emscripten support)

Also changes all files which reference SDL or GL to use a separate
APG-provided import. This ensures that OS-specific changes are easy to
make (e.g. on MacOS or for Emscripten)

There's also a workaround which disables SDL_mixer version logging since that
function is marked TODO in the emscripten port.